pbs-353--refactor: gpt 도메인 설계
