+++ backend/src/main/java/com/pobluesky/backend/domain/chat/controller/ChatController.java
@@ -1,0 +1,14 @@
+package com.pobluesky.backend.domain.chat.controller;
+
+import lombok.RequiredArgsConstructor;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/api/chat")
+public class ChatController {
+
+
+}
