+++ backend/src/main/java/com/pobluesky/backend/domain/chat/service/OcrService.java
@@ -1,4 +1,4 @@
-package com.pobluesky.backend.domain.inquiry.service;
+package com.pobluesky.backend.domain.chat.service;
 
 import com.google.cloud.vision.v1.*;
 import com.pobluesky.backend.global.error.CommonException;
